# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/profile/me": {
      "get": {
        "operationId": "ProfileController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Мой профиль",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/me/collections": {
      "get": {
        "description": "Имя коллекции — то же значение, что принимает POST /posts/{id}/favorite в поле `collection`, поэтому список можно показывать выбором вместо ввода руками. Посты вне коллекций сюда не входят — они в /profile/favorites.",
        "operationId": "ProfileController_collections",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CollectionDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Мои коллекции сохранённого",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/favorites": {
      "get": {
        "operationId": "ProfileController_favorites",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Курсор последнего элемента предыдущей страницы",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostBriefDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Сохранённое (только своё)",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/me/reposts": {
      "get": {
        "operationId": "ProfileController_reposts",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Курсор последнего элемента предыдущей страницы",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostBriefDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Мои репосты",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/me/saved-music": {
      "get": {
        "operationId": "ProfileController_savedMusic",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MusicBriefDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Сохранённая музыка",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/me/activity": {
      "get": {
        "description": "Лайки, комментарии, просмотры и поисковые запросы одним списком, с фильтром по дате.",
        "operationId": "ProfileController_activity",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Курсор последнего элемента предыдущей страницы",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Фильтр: не раньше этой даты",
            "schema": {
              "example": "2026-07-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Фильтр: не позже этой даты",
            "schema": {
              "example": "2026-07-31",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityItemDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ваши действия",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile": {
      "put": {
        "description": "about ≤150, website, gender (симметричный enum), occupation, dob, флаги.",
        "operationId": "ProfileController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Изменить профиль",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/privacy": {
      "put": {
        "operationId": "ProfileController_privacy",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePrivacyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Закрытый аккаунт вкл/выкл",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/avatar": {
      "put": {
        "operationId": "ProfileController_setAvatar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Загрузить аватар",
        "tags": [
          "profile"
        ]
      },
      "delete": {
        "description": "Обнуляет только Profile.avatarUrl. Логин НЕ ломается (баг softclub #2).",
        "operationId": "ProfileController_deleteAvatar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvatarDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Удалить аватар",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/{userId}": {
      "get": {
        "description": "+ isFollowing / isFollowedBy / isBlocked / hasRequestPending / canViewContent",
        "operationId": "ProfileController_byId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtherProfileDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Профиль пользователя",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/{userId}/is-following": {
      "get": {
        "operationId": "ProfileController_isFollowing",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsFollowingDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Подписан ли я на этого пользователя",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/{userId}/posts": {
      "get": {
        "operationId": "ProfileController_posts",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Курсор последнего элемента предыдущей страницы",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostBriefDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Публикации пользователя (закрытый аккаунт → 403)",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/{userId}/reels": {
      "get": {
        "operationId": "ProfileController_reels",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Курсор последнего элемента предыдущей страницы",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostBriefDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reels пользователя (закрытый аккаунт → 403)",
        "tags": [
          "profile"
        ]
      }
    },
    "/api/profile/{userId}/tagged": {
      "get": {
        "operationId": "ProfileController_tagged",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Курсор последнего элемента предыдущей страницы",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PostBriefDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Отмеченные публикации (закрытый аккаунт → 403)",
        "tags": [
          "profile"
        ]
      }
    },
...
  },
  "info": {
    "title": "Instagram Backend API",
    "description": "NestJS + Prisma + PostgreSQL. Конверт ответа: { data, errors, statusCode }",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "ProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userName": {
            "type": "string",
            "example": "eraj"
          },
          "fullName": {
            "type": "string",
            "example": "Eraj Karimov"
          },
          "avatarUrl": {
            "type": "object",
            "nullable": true
          },
          "about": {
            "type": "object",
            "nullable": true,
            "maxLength": 150
          },
          "website": {
            "type": "object",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "HIDDEN"
            ],
            "example": "HIDDEN"
          },
          "occupation": {
            "type": "object",
            "nullable": true
          },
          "dob": {
            "type": "object",
            "nullable": true
          },
          "showThreadsBadge": {
            "type": "boolean",
            "example": false
          },
          "isAiAuthor": {
            "type": "boolean",
            "example": false
          },
          "showAccountSuggestions": {
            "type": "boolean",
            "example": true
          },
          "isPrivate": {
            "type": "boolean",
            "example": false
          },
          "isVerified": {
            "type": "boolean",
            "example": false
          },
          "postsCount": {
            "type": "number",
            "example": 42
          },
          "followersCount": {
            "type": "number",
            "example": 128
          },
          "followingCount": {
            "type": "number",
            "example": 97
          }
        },
        "required": [
          "id",
          "userName",
          "fullName",
          "gender",
          "showThreadsBadge",
          "isAiAuthor",
          "showAccountSuggestions",
          "isPrivate",
          "isVerified",
          "postsCount",
          "followersCount",
          "followingCount"
        ]
      },
      "CollectionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Путешествия",
            "description": "Имя коллекции — оно же ключ при сохранении"
          },
          "postsCount": {
            "type": "number",
            "example": 7
          },
          "coverUrl": {
            "type": "object",
            "nullable": true,
            "description": "Превью последнего сохранённого поста (или явная обложка коллекции)"
          }
        },
        "required": [
          "name",
          "postsCount"
        ]
      },
      "PostBriefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 12
          },
          "caption": {
            "type": "object",
            "nullable": true
          },
          "isReel": {
            "type": "boolean",
            "example": false
          },
          "coverUrl": {
            "type": "object",
            "nullable": true,
            "description": "Первое медиа — обложка в сетке"
          },
          "likesCount": {
            "type": "number",
            "example": 24
          },
          "commentsCount": {
            "type": "number",
            "example": 3
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "isReel",
          "likesCount",
          "commentsCount",
          "createdAt"
        ]
      },
      "MusicBriefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 7
          },
          "title": {
            "type": "string",
            "example": "Blinding Lights"
          },
          "artist": {
            "type": "string",
            "example": "The Weeknd"
          },
          "coverUrl": {
            "type": "object",
            "nullable": true
          },
          "savedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "artist",
          "savedAt"
        ]
      },
      "ActivityItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "LIKE:12",
            "description": "Составной id «ТИП:id строки». Список слит из четырёх таблиц, где id повторяются, поэтому простого id тут не существует. Для «показать ещё» используйте `at` (см. cursor)."
          },
          "type": {
            "type": "string",
            "enum": [
              "LIKE",
              "COMMENT",
              "POST_VIEW",
              "SEARCH"
            ],
            "example": "LIKE"
          },
          "at": {
            "format": "date-time",
            "type": "string",
            "example": "2026-07-15T10:00:00.000Z"
          },
          "postId": {
            "type": "number",
            "example": 12,
            "description": "id поста — для LIKE / COMMENT / POST_VIEW"
          },
          "text": {
            "type": "string",
            "description": "Текст — для COMMENT и SEARCH"
          }
        },
        "required": [
          "id",
          "type",
          "at"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "about": {
            "type": "string",
            "example": "Фотограф из Душанбе",
            "maxLength": 150
          },
          "website": {
            "type": "string",
            "example": "https://eraj.dev"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "HIDDEN"
            ],
            "example": "MALE",
            "description": "Симметричный enum: что отправили — то и вернётся (баг softclub #12)"
          },
          "occupation": {
            "type": "string",
            "example": "Фотограф"
          },
          "dob": {
            "type": "string",
            "example": "2000-05-17"
          },
          "locationId": {
            "type": "number",
            "example": 1,
            "description": "id локации"
          },
          "showThreadsBadge": {
            "type": "boolean",
            "example": false
          },
          "isAiAuthor": {
            "type": "boolean",
            "example": false
          },
          "showAccountSuggestions": {
            "type": "boolean",
            "example": true
          },
          "fullName": {
            "type": "string",
            "example": "Eraj Karimov",
            "description": "Имя и фамилия"
          }
        }
      },
      "UpdatePrivacyDto": {
        "type": "object",
        "properties": {
          "isPrivate": {
            "type": "boolean",
            "example": true,
            "description": "true — закрытый аккаунт"
          }
        },
        "required": [
          "isPrivate"
        ]
      },
      "AvatarDto": {
        "type": "object",
        "properties": {
          "avatarUrl": {
            "type": "object",
            "nullable": true,
            "description": "null после удаления"
          }
        }
      },
      "OtherProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userName": {
            "type": "string",
            "example": "eraj"
          },
          "fullName": {
            "type": "string",
            "example": "Eraj Karimov"
          },
          "avatarUrl": {
            "type": "object",
            "nullable": true
          },
          "about": {
            "type": "object",
            "nullable": true,
            "maxLength": 150
          },
          "website": {
            "type": "object",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "HIDDEN"
            ],
            "example": "HIDDEN"
          },
          "occupation": {
            "type": "object",
            "nullable": true
          },
          "dob": {
            "type": "object",
            "nullable": true
          },
          "showThreadsBadge": {
            "type": "boolean",
            "example": false
          },
          "isAiAuthor": {
            "type": "boolean",
            "example": false
          },
          "showAccountSuggestions": {
            "type": "boolean",
            "example": true
          },
          "isPrivate": {
            "type": "boolean",
            "example": false
          },
          "isVerified": {
            "type": "boolean",
            "example": false
          },
          "postsCount": {
            "type": "number",
            "example": 42
          },
          "followersCount": {
            "type": "number",
            "example": 128
          },
          "followingCount": {
            "type": "number",
            "example": 97
          },
          "isFollowing": {
            "type": "boolean",
            "example": false,
            "description": "Я подписан на него"
          },
          "isFollowedBy": {
            "type": "boolean",
            "example": false,
            "description": "Он подписан на меня"
          },
          "isBlocked": {
            "type": "boolean",
            "example": false,
            "description": "Я его заблокировал"
          },
          "hasRequestPending": {
            "type": "boolean",
            "example": false,
            "description": "Моя заявка на подписку ждёт подтверждения"
          },
          "canViewContent": {
            "type": "boolean",
            "example": true,
            "description": "Виден ли контент: у закрытого аккаунта — только принятым подписчикам"
          }
        },
        "required": [
          "id",
          "userName",
          "fullName",
          "gender",
          "showThreadsBadge",
          "isAiAuthor",
          "showAccountSuggestions",
          "isPrivate",
          "isVerified",
          "postsCount",
          "followersCount",
          "followingCount",
          "isFollowing",
          "isFollowedBy",
          "isBlocked",
          "hasRequestPending",
          "canViewContent"
        ]
      },
      "IsFollowingDto": {
        "type": "object",
        "properties": {
          "isFollowing": {
            "type": "boolean",
            "example": true
          },
          "hasRequestPending": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "isFollowing",
          "hasRequestPending"
        ]
      },
...
    }
  }
}